fixes race condition in observer
